elfutils: build out of tree on libc++
`elfutils` has a `stack` binary built in `src`.

When built in source tree this breaks when `#include <stack>` is
transitively included in
`libc++` via `<iostream>`.

The trigger set of options is `-I $(srcdir)` which turns to `-I .` with
`./stack` present.

The fix turns the above to `-I ../src` without `./stack` present.

The change is conditional on `LLVM` to avoid `stdenv` rebuild. We will
turn it into unconditional in `staging` once this gets merged.

While at it dropped `CFLAGS` / `CXXFLAGS` overrides.