workflows/labels: label stale issues
By re-organizing the flow in `handle()` we can start labeling both
issues and pull requests, and only make the relevant API requests for
the PR-case.

At first glance, we might think that we only need to label the big batch
list of issues and not those recently updated: But that's wrong, for
recently updated issues it's important to label quickly, because the
stale label needs to be *removed*, too.

(cherry picked from commit 2fa1151e544454140bf72b5304c7b443dd6fb5ca)