catalyst-browser: drop
catalyst is electron browser which are unsafe be design due to lack of
proper sandboxing.

The package also seems unmaintained upstream.
Updates were not backported too, which is not accaptable for a browser.

Also, it has no commiter among its maintainers, which [is required](https://github.com/NixOS/nixpkgs/tree/master/pkgs#quick-start-to-adding-a-package).