cc-wrapper: use 64‐bit file and time types on 32‐bit platforms
This is going to be such a great commit message when it’s
written. Seriously you’re hardly going to believe it. It’ll be
like, wow, this is such a good idea that I’m hitting the merge
button immediately. Retroactively, even. By the time you are reading
the commit message you will have already merged it into Nixpkgs
forever. It’ll be 2038 and everything will be great and it’ll
all be thanks to you.