nixos/sssd: add upstream directives in sssd.service
Changed the service type from forking to notify,
which should gives a better indication of when the service is ready.

Changed the preStart into an ExecStart,
in order for upstream's NotifyAccess=main to work.

Added Restart=on-abnormal for better service stability.

(cherry picked from commit 98672296968c1852754a6bb9993242420613e02c)