nixos/gitea: drop mailerUseSendmail option and use PROTOCOL as an indication instead
This option does not configure sendmail itself because it is impossible
as sendmail is an alias for many things and could mean msmtp or postfix
or exim or something else.

Instead we rely on the PROTOCOL setting as initially proposed #384582
and based on that open up the sandboxing settings because if the user
configures sendmail, they want it to work and not have to configure yet
another things.

Also makes postfix specific things conditional on postfix being enabled
as msmtp does not need them.

Also we can set SENDMAIL_PATH unconditionally as every wrapper I am
aware of uses that path.