nixos-rebuild-ng: refactor part of __init__.py code in services.py
There are a bunch of logic that were happening in `__init__.py` file
that was not really fit in other files in the project but also didn't
seems fit to the `__init__.py`.

While moving this to `services.py` doesn't really solve the problem, at
least it reduces the size of the main `execute()` function to a
reasonable size, and I hope this will make it easier for other people to
understand what is happening.