nushellPlugins.*: add `versionCheckHook`
Why for each all at the same time? Because all nushell plugins should
respond in the same way with either no arguments passed or with
`--help`, which contains their version.

Also because when new plugins get added or plugins get updated, problems
can be easily spotted because the errors are really loud.