nixos/nix-optimise: run with lower priority
It's painfully obvious to me on my anaemic system when `nix-optimise` kicks in,
so run it with much lower priority since it's not time critical anyway.

Additionally, do not run on battery power.

I can appreciate that these are oppinionated defaults but I am having a hard
imagining cases where this would be a show-stopper so there isn't much point in
making this directly configurable via the nixos module system as people can
always override the unit directly themselves via `systemd.services.nix-optimise`