nitter: make about page's version string match upstream
The /about page in upstream uses the format "YYYY.MM.DD-<short commit
hash>"
in the footer. However, the nix package currently uses the nix version
string
for reproducibility. This causes issues with the [nitter instance status
page](https://status.d420.de)
not being able to detect the version of any instance running with nix,
so
this commit tries to make it match what the impure upstream version
would do.