libfdt: Rename the package to pylibfdt
The “canonical” name of the Python libfdt package, i.e. the one used on
PyPI and, hence, the one that is listed in the requirements of the
dependent modules is `pylibfdt`, but the upstream build system produced
a package called `libfdt`.

Given that, realistically, everyone will be using `pylibfdt`, as seen on
PyPI, our choices are to either patch / jailbreak the requirement in all
dependent packages, or follow PyPI and rename the package itself.

Rename the package once and for all, since it is easier than modifying
every single dependent one.