buildGoModule: do not emit broken position-independent executables
Unless cgo is enabled, the Go compiler uses a custom linker. PIEs built by this
linker always use the dynamic loader, which defaults to the FHS path. Disable
position-independent executables when static linkage is desired so that the
output is actually statically linked. In addition try to specify the correct
dynamic loader in the other cases.