fuse: fix build
configure.ac uses the AM_ICONV macro, which is part of gettext.  The
patch makes autoreconf recognise the project as using gettext, which
has the added benefit of meaning we no longer need to run makeconf.sh,
and can rely on just the autoreconf from autoreconfHook.

Previously, we were actually doing autoreconf twice; once from the
hook and once from the script, which I doubt was intentional…