rustc: Set rust.frame-pointers if !isx86_32
We want -Cforce-frame-pointers=yes when building rustc itself, but the
bootstrap process does not use cargoSetupHook, so set the bootstrap
config option rust.frame-pointers = true if target is not x86_32.

Note that std is unconditionally compiled with frame pointers. See code
comment for elaboration.

See also #399014 for discussion when we enabled frame pointers for C.