pkgs/README.md: discourage overrideAttrs within Nixpkgs
This comes up quite a lot in review, and we've just had a close call
with a known vulnerable library version almost being reintroduced
without appropriate knownVulnerabilities markings in an override in
the expression for a leaf package.  Let's take the first step towards
preventing that sort of thing happening again.

Co-authored-by: Robert Hensing <roberth@users.noreply.github.com>