xen: patch with XSA-470
https://xenbits.xenproject.org/xsa/advisory-470.html

Certain instructions need intercepting and emulating by Xen.  In some
cases Xen emulates the instruction by replaying it, using an executable
stub.  Some instructions may raise an exception, which is supposed to be
handled gracefully.  Certain replayed instructions have additional logic
to set up and recover the changes to the arithmetic flags.

For replayed instructions where the flags recovery logic is used, the
metadata for exception handling was incorrect, preventing Xen from
handling the the exception gracefully, treating it as fatal instead.

An unprivileged guest can cause a hypervisor crash, causing a Denial of
Service (DoS) of the entire host.

Signed-off-by: Fernando Rodrigues <alpha@sigmasquadron.net>