cloud-init: 24.2 -> 25.1.3
Auto update broke with version 24.3, this was caused by the following
commit: canonical/cloud-init@c28092f

The above commit changed some tests to write temporary files with
specific permissions, which seems to be causing trouble for us. I've
included a patch which removes these permissions, and the build is
working fine now.