buildstream: 2.4.1 -> 2.5.0
I've had to override the `protobuf` input for `python3Packages.grpcio`,
`python3Packages.protobuf`, and `buildbox` to `protobouf_27`, so that it
works with the outdated Protobuf requirement Buildstream currently
supports.

For now, it is important to get 2.5.0 landed, as it includes a new API
which is necessary for new releases of Buildstream plugins.