kubernetes: use util-linuxMinimal instead of util-linux.withPatches
This, and commits to k3s and util-linux close #409339.
The util-linux.withPatches API was a temporary hack for the 25.05
release to fix Kubernetes, and is going away.

While we're at it, we should use util-linuxMinimal because we do not
need things such as systemd support for kubelet initialization.

(cherry picked from commit 949e299d246a7e2d64402dec4c3c0a841416987a)