grafana: change convention on Go version changes
Grafana sets the min version to the latest Go patch-level, but on ~every
other release it happens that we're a little bit behind and then the
build fails, so I don't remove the technically dead-code just to pull it
out of history next time.

The problem is that stable may be a little behind with Go and then the
backport will fail. So from now on we just use the `.0` release of the
minor we need to prevent that.

If the minor isn't on stable yet, we need a manual backport anyways.

(cherry picked from commit dd4e9bdc17dd9fd88815008f36348099dd6c3dd1)