nixos/ollama: Fix race condition with AMD GPU driver on boot
Fixes #419097

It was possible for the Ollama service to load before the AMD GPU driver had
finished loading on boot.
As a result the `/sys/class/kfd/kfd/topology/nodes/*/properties` path didn't
exist and Ollama would fallback to execution on the CPU.
This change also requires users to have `boot.kernelModules = [ "amdgpu" ];`
in their configuration to ensure that `/etc/modules.d/nixos.conf` is
populated and `systemd-modules-load.service` waits on the driver.