ci: increase limit of PR reviewers and assign first few if too many
Currently, if too many PR reviewers are detected, then none of them are
assigned. This can easily happen if two mid-size teams are maintaining a
single package.

This change is increasing the limit of reviewers to 15, which looks to
work with GiHub, and in case it is still not enough it will assign first
15 people as reviewers instead of doing nothing.