odin: use external Raylib instead of vendored
The default behavior is to use the statically linked Raylib libraries, but GLFW still attempts to load Xlib at runtime, which won't normally be available on Nix based systems. Instead, use the "system" Raylib version, which can be provided by a pure Nix expression, for example in a shell.

The alternative solution would be to either
- replace the vendored libraries with ones from the nixpkgs Raylib derivation
- patch the vendored libraries with the required dependencies
Both of these solutions require making the Odin package depend on libraries that it shouldn't.