Patch to allow Workrave GNOME extension to work
This addresses Issue #420671. It both patches the JavaScript in the GNOME extension itself, as well as patches to the C and C++ source so that Workrave's GSettings schemas can be found by the parts of Workrave that are used in the extension and are not affected by the wrapper that allows the main Workrave app to function.

This is for the extension for GNOME versions 45 and up.