sonic-pi: add qtwayland dep
Running sonic-pi under wayland/sway, I noticed the following output:

```
qt.qpa.plugin: Could not find the Qt platform plugin "wayland" in ""
```

I couldn't find any functional issue due to this, but looked into it
anyway.

Adding `qtwayland` dependency resolves this issue, it now prints:

```
qt.qpa.wayland: Wayland does not support QWindow::requestActivate()

```

The dependency on `qtwayland` is documented upstream in
https://github.com/sonic-pi-net/sonic-pi/blob/5145973b25a7f21cf94151f5978ba4819550012d/BUILD-LINUX.md?plain=1#L63

Which looks a bit healthier. Not sure if there are other things to
improve for wayland support, but this is a start.

Signed-off-by: Paul Meyer <katexochen0@gmail.com>