pylibfdt: Use the correct Python for tests
By default, Meson will find the version of Python that it itself is
using, i.e. the default one in Nixpkgs.
However, when building libfdt for other Python package sets, we need to
make sure we match the version of the Python used for tests – otherwise
the interpreter will not find the compiled module.

Tweak meson.build to always use the Python interpreter that was passed
to the expression, rather than the default one.