stdenv: Don't reappend patchFlags
Presently, each patch in the for loop for patchPhase re-appends the
patchFlags. This can be harmless "patch -p1 -p1" is effectively the same
as "patch -p1". Other flags to patch, such as "-d" perform action each
time it is present in the argument list. The flagsArray doesn't contain
any patch-specific arguments and can be declared outside the loop.