materia-kde-theme: update to 20220823-unstable-2023-07-15
The latest available release tag is `20220823`, but there was one more
commit added on 2023-07-15. The project doesn't get updated anymore, but
with some fiddling around with *qt5ct* or *qt6ct*, it can be made to
work.

Since the last commit was two years ago, I think it's worth setting the
version to something funny like `20220823-unstable-2023-07-15` for the
sake of sneaking in that one last commit.