nixos/tests/meilisearch: fix race condition when creating index
previously, when creating the index, there is a race condition between
the task being created and when we check if the index was successfully
created. if it is created too slow, the test fails.

it already polled/awaited when adding documents, so i generalized this
logic (and made it terminate anytime the task finishes, even if that
wasn't success). not ideal, since meilisearch supports proper webhook
callbacks, but whatever, we're writing tests, not production code.