nixos/openafs: Extend CellServDB options
Previously, servers could only be supplied for ThisCell. This meant that
if a server and client have different ThisCell values, the client's
CellServDB couldn't be configured for its own server.

The previous list values are still accepted, but now a CellServDB option
can also be specified as an attrset of cell names to servers.