podman-desktop: force disable electron-builder signing on arm64
Even with CSC_IDENTITY_AUTO_DISCOVERY=false, latest builder attempts to
sign if on arm64 unless identity is forcefully set to `null`.

The problem that builder tries to solve with falling back to an ad-hoc
signature doesn't affect us because we later sign binaries with our own
hook. So it should be safe to disable builder's signature logic.