python3Packages.docutils: Fix upstream URL (repo.or.cz -> sourceforge)
I am seeing frequent HTTP 503 Service Unavailable responses from
repo.or.cz, and I'm looking for an alternative upstream.

The project's ultimate upstream appears to be SourceForge: the project
homepage is hosted by SourceForge and the README references SourceForge
as the ultimate source of releases.

https://docutils.sourceforge.io/docs/dev/repository.html mentions the
Git mirror at repo.or.cz, but this appears to be more of a (read-only?)
convenience for project developers that prefer Git. There is also this
footnote:

  There are also 3rd-party mirrors and forks at GitHub, some of them
  orphaned. At the time of this writing (2021-11-03),
  https://github.com/live-clones/docutils/tree/master/docutils provides
  an hourly updated clone.

but this clone does not contain release tags.

Due to how this project is organized and the connectivity problems to
repo.or.cz, I suspect that the releases hosted on SourceForge are a
better upstream for nixpkgs, than to download corresponding releases
from the Git mirror at repo.or.cz.