python3Packages.buildPythonPackage: default enabledTestPaths = [ "." ]
Explicitly provide a default value for the argument `enabledTestPaths`
as the current working directory (`enabledTestPaths = [ "." ]`)
to simplify overriding (making `enabledTestPaths` always specified as a
non-empty list) while preserving `pytest`'s default behaviour of
discovering tests in the current working directory.

Assert the specified `enabledTestPaths` value to be a non-empty list.