python3Packages.pytz: use system tzdata
We update that one more reliably, and it allows system-level
customizations to apply to pytz as well. This is in line with the
zoneinfo stdlib module in Python >=v3.9 which accesses the system
timezone database directly.

This is modeled on how the certifi module uses the system-wide
ca-bundle, see commit 8d7cc9cac9ecdf95f554c5ea7ca15118baa06c39.