pinentry-kwallet: init at 3.03
pinentry-kwallet is a pinentry program for GnuPG that uses KWallet to store passphrases. This allows automatically unlocking GnuPG keys on login. This package is a thin wrapper that symlinks to the pinentry-kwallet binary in the kwalletcli package to make it compatible with the `programs.gnupg.agent.pinentryPackage` option in NixOS.