.github/PULL_REQUEST_TEMPLATE: avoid inlining links for readability
We expect the TODO list to be read through *on PR creation*, otherwise
the html comments would not make sense. Thus, we should make it even
only slightly readable, which was not at all the case before.

The links for release notes are removed, because the PR author has no
value from the *current* release notes. They will need to find the file
manually anyway.