.github/PULL_REQUEST_TEMPLATE: shorten inline comments
The shorter the better, because they take space - and won't be read if
too long.

The notes about the changelog or the upstream homepage are important..
but are too late to add here. They should *already* be in the commit
message - and when they are, they will be in the PR template
automatically. On a single commit PR, at least.

We'll somehow need to find a way to motivate authors to understand the
contribution guidelines *first*, before they open the PR - so that they
already have the changelog sorted out.

Also removed the duplicate "as applicable" from the Tested section. The
notion of "only when applicable" is already communicated in the inline
comment above.