workflows/build: build x86_64-darwin shell via rosetta
Using the older macos-13 runners induces significant queue times before
the job even starts - and then the runners are really slow. This
consistently makes the x86_64-darwin build job the slowest of the pack.

By running this on a macos-14 runner, which is arm64, this becomes
faster - despite using rosetta for emulation.